backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
PyMySQL==1.1.0
PyJWT[crypto]==2.8.0
cryptography==42.0.5
requests==2.31.0
python-dotenv==1.0.1